Split patch code definitions in separate headers
