+++ arm9/source/patches/platform/supercard/sclite/SuperCardLiteReadSectorPatchCode.s
