+++ arm9/source/patches/platform/supercard/sclite/SuperCardLiteSendCommandPatchCode.s
