+++ arm9/source/patches/platform/supercard/sclite/SuperCardLiteWriteSectorPatchCode.s
