+++ arm9/source/patches/platform/supercard/scsd/SuperCardSDReadSectorPatchCode.s
