+++ arm9/source/patches/platform/supercard/scsd/SuperCardSDSendCommandPatchCode.s
