+++ arm9/source/patches/platform/supercard/scsd/SuperCardSDWriteSectorPatchCode.s
